Moved view to its own folder, added utility
