+++ src/app/app.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {Router, Link} from "@reach/router"
 
-import CreateChallenge from '../views/createChallenge'
+import CreateChallenge from '../views/createChallenge/createChallenge'
 import View from '../components/view/view'
 import {
   createChallenge,
